Now possible to get the time series ready for compression and save the compressed data
